리액션 카운트, 순서 리얼타임 업데이트 기능 구현, 순위 문제도 해결 완료
